+++ BZI/sampling.py
@@ -171,21 +171,22 @@ def sphere_pts(A, r2, offset=[0.,0.,0.], eps=1e-9):
             coordinates.
     """
 
-    # This is a parameter that should help deal with rounding error.
     offset = np.asarray(offset)
     
-    # Put the offset in cell coordinates
-    oi= np.round(np.dot(inv(A),offset))
-    # Find a cell point close to the offset
-    oi = oi.astype(int)
+    # Put the offset in cell coordinates and find a cell point close to the
+    # offset.
+    oi= np.round(np.dot(inv(A),offset)).astype(int)
     r = np.sqrt(r2)
     V = np.linalg.det(A)
     n = [int(np.ceil(norm(np.cross(A[:,(i+1)%3],A[:,(i+2)%3]))*r/V) + 1)
-         for i in range(3)]         
-    ints = product(range(-n[0] + oi[0], n[0] + oi[0]),
-                   range(-n[1] + oi[1], n[1] + oi[1]),
-                   range(-n[2] + oi[2], n[2] + oi[2]))
-    grid = np.array([np.dot(A, comb) for comb in ints])
+         for i in range(3)]
+
+    ints = np.array(list(product(range(-n[0] + oi[0], n[0] + oi[0] + 1),
+                   range(-n[1] + oi[1], n[1] + oi[1] + 1),
+                   range(-n[2] + oi[2], n[2] + oi[2] + 1))))
+    
+    grid = np.dot(A, ints.T).T - offset
+    # grid = np.array([np.dot(A, comb) - offset for comb in ints])
     norms = np.array([np.dot(p,p) for p in grid])
     return grid[np.where(norms < (r2 + eps))]